pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ci-cd-supply-chain-auditor"
version = "1.0.0"
description = "A security-focused auditor for CI/CD pipelines, identifying vulnerabilities and misconfigurations in your software supply chain."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Raghu", email = "raghu@example.com"}
]
keywords = [
    "security",
    "ci-cd",
    "supply-chain",
    "audit",
    "github-actions",
    "slsa",
    "devsecops"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1.0,<9.0.0",
    "PyYAML>=6.0.1,<7.0.0",
    "gitpython>=3.1.40,<4.0.0",
    "requests>=2.31.0,<3.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "jinja2>=3.1.2,<4.0.0",
    "rich>=13.7.0,<14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "ruff>=0.1.9",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.31.0",
]

[project.scripts]
auditor = "auditor.cli:main"

[project.urls]
Homepage = "https://github.com/raghu-007/CI-CD-Supply-Chain-Auditor"
Documentation = "https://github.com/raghu-007/CI-CD-Supply-Chain-Auditor#readme"
Repository = "https://github.com/raghu-007/CI-CD-Supply-Chain-Auditor.git"
Issues = "https://github.com/raghu-007/CI-CD-Supply-Chain-Auditor/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["auditor*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["git.*", "yaml.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "S",      # flake8-bandit (security)
    "T20",    # flake8-print
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # assert usage (needed for tests)
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "S105", "S106"]

[tool.coverage.run]
source = ["auditor"]
branch = true
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
